tests/wpt/tests/html/dom/render-blocking/element-render-blocking-018.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Media attribute changes in the body to apply, but has no effect</title>

<link id=link rel=expect href="#last" blocking="render" media="(max-width: 10px)">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "changing media to matching in the body has no effect");
</script>
</head>
<body>
<script>
link.media = "(min-width: 10px)";
</script>

  <div id="first"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="second"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="last"></div>
</body>
